DE1514 Route refactor Issues
* merged adminTool.routes.js and mpTools.routes.js
* re-added thedaily route
