+++ crossroads.net/app/routes.js
@@ -508,7 +508,40 @@
               }).$promise;
             }
           }
+        })
+				.state('thedaily', {
+          parent: 'noSideBar',
+          url: '/thedaily',
+          controller: 'TheDailyController as dailyCtrl',
+          templateUrl: 'thedaily/thedaily.html'
         });
+
+        
+  			// Commented out for US2924, will be added back after Corkboard go-live
+				//
+				//.state('corkboard', {
+				//  url: '{link:corkboardRouteType}',
+				//  resolve: {
+				//    RedirectToSubSite: function($window, $location) {
+				//      // Force browser to do a full reload to load corkboard's index.html
+				//      $window.location.href = $location.path();
+				//    }
+				//  },
+				//  data: {
+				//    preventRouteAuthentication: true,
+				//    meta: {
+				//      title: 'Corkboard',
+				//      description: ''
+				//    }
+				//  }
+				//})
+
+        //Leave the comment below.  Once we have a true 404 page hosted in the same domain, this is how we
+        //will handle the routing.
+        //.state('404', {
+        //    templateUrl: __CMS_ENDPOINT__ + '/page-not-found/'
+        //}); 
+        ;
         $urlRouterProvider.otherwise('/');
   }
